認証制御をinteractorsに移動し処理を修正
