+++ apps/web/controllers/explorer/execute.rb
@@ -1,7 +1,7 @@
 module Web::Controllers::Explorer
   class Execute
     include Web::Action
-    include Web::Controllers::Concerns::AuthFilter
+    include UserInteractor::Util
 
     before :authenticate!
     expose :endpoints, :result
