+++ apps/web/controllers/explorer/index.rb
@@ -1,10 +1,10 @@
 module Web::Controllers::Explorer
   class Index
     include Web::Action
-    include Web::Controllers::Concerns::AuthFilter
+    include UserInteractor::Util
 
     before :authenticate!
-    expose :endpoints
+    expose :endpoints, :result
 
     def call(params)
       @endpoints = Endpoint.all.each_with_object({}){|(target, methods), obj|
