+++ apps/web/templates/explorer/index.html.erb
@@ -35,4 +35,4 @@
 <hr>
 
 <h2>実行結果</h2>
-<div id="result"></div>
+<div id="result"><%= h result %></div>
