+++ apps/web/views/explorer/execute.rb
@@ -1,5 +1,7 @@
 module Web::Views::Explorer
   class Execute
     include Web::View
+
+    template 'explorer/index'
   end
 end
